requirements/base.txt
Unipath

bootstrap-admin

django-bootstrap3

requests

django-nested-formset

django-nested-inline